+++ src/CastTo/Entity.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace Nandan108\DtoToolkit\CastTo;
 
-use Nandan108\DtoToolkit\Contracts\CreatesFromArrayOrEntityInterface;
 use Nandan108\DtoToolkit\Core\BaseDto;
 use Nandan108\DtoToolkit\Core\CastBase;
 use Nandan108\DtoToolkit\Exception\Process\TransformException;
@@ -12,7 +11,7 @@ use Nandan108\DtoToolkit\Internal\Exporter;
 
 /**
  * Casts arrays or nested DTO to a specified Entity class
- * If input is array, will use CreatesFromArrayOrEntityInterface::newFromArray.
+ * Actual export is delegated to Exporter::export method.
  *
  * @psalm-suppress UnusedClass
  */
@@ -20,17 +19,25 @@ use Nandan108\DtoToolkit\Internal\Exporter;
 final class Entity extends CastBase
 {
     /**
-     * @param ?class-string $entityClass
+     * @param ?class-string $entityClass       The entity class to use when exporting. Resolution order:
+     *                                         1. If $entityClass param is provided, it is used
+     *                                         2. If $value is a DTO with #[DefaultOutboundEntity] attribute, with matching scope, that class is used.
+     *                                         3. If $value implements PreparesEntityInterface, it is used to instantiate the entity.
+     *                                         4. Otherwise, InvalidConfigException is thrown.
+     * @param array         $supplementalProps Additional data to set on the entity. This can be used to set
+     *                                         relations or other properties that are not part of the DTO.
+     *                                         These properties are not subject to #[MapTo] attributes.
+     * @param bool          $recursive         whether to recursively convert nested DTOs to entities
      */
-    public function __construct(?string $entityClass = null, array $extraProps = [], bool $recursive = false)
+    public function __construct(?string $entityClass = null, array $supplementalProps = [], bool $recursive = false)
     {
-        parent::__construct([$entityClass, $extraProps, $recursive]);
+        parent::__construct([$entityClass, $supplementalProps, $recursive]);
     }
 
     #[\Override]
     public function cast(mixed $value, array $args): mixed
     {
-        if (!(is_array($value) || $value instanceof BaseDto)) {
+        if (!(\is_array($value) || $value instanceof BaseDto)) {
             throw TransformException::expected(
                 operand: $value,
                 methodOrClass: self::class,
@@ -38,17 +45,12 @@ final class Entity extends CastBase
             );
         }
 
-        /** @var class-string|null $explicitEntityClass */
-        /** @var bool $recursive */
-        /** @var array $extraProps */
-        [$explicitEntityClass, $extraProps, $recursive] = $args;
-
         return Exporter::export(
             source: $value,
             as: 'entity',
-            entity: $explicitEntityClass,
-            extraProps: $extraProps,
-            recursive: $recursive,
+            entity: $args[0],
+            supplementalProps: $args[1],
+            recursive: $args[2],
         );
     }
 }