+++ src/Traits/ExportsOutbound.php
@@ -18,16 +18,16 @@ trait ExportsOutbound
     /**
      * Convert the DTO to an entity.
      *
-     * @param object|class-string|null $entity     The entity to fill. If null, a new instance will be created.
-     * @param array                    $extraProps Additional data to set on the entity. This can be used to set
-     *                                             relations or other properties that are not part of the DTO.
-     * @param bool                     $recursive  whether to recursively convert nested DTOs to entities
+     * @param object|class-string|null $entity            The entity to fill. If null, a new instance will be created.
+     * @param array                    $supplementalProps Additional data to set on the entity. This can be used to set
+     *                                                    relations or other properties that are not part of the DTO.
+     * @param bool                     $recursive         whether to recursively convert nested DTOs to entities
      *
      * @throws InvalidConfigException
      */
     public function exportToEntity(
         string | object | null $entity = null,
-        array $extraProps = [],
+        array $supplementalProps = [],
         ?ProcessingErrorList $errorList = null,
         ?ErrorMode $errorMode = null,
         bool $recursive = false,
@@ -38,7 +38,7 @@ trait ExportsOutbound
             as: 'entity',
             entity: $entity,
             errorList: $errorList,
-            extraProps: $extraProps,
+            supplementalProps: $supplementalProps,
             errorMode: $errorMode,
             recursive: $recursive,
         );